% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.extension.R
\name{.add.extension}
\alias{.add.extension}
\title{Associate a reader function with an extension.}
\usage{
.add.extension(extension, reader)
}
\arguments{
\item{extension}{The extension of the new data file.}

\item{reader}{The function to use when reading the data file. It MUST accept
three arguments: \code{filename}, \code{variable.name} and \code{...} (in
that order and called by those names).  The function should read the
contents of the file \code{filename}, and save it into the workspace under
the name \code{variable.name}.  The \code{...} argument is just an optional
interface for the \code{file.reader} to pass in extra options.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will associate an extension with a custom reader function.
}
\section{Warning}{

This interface should not be considered as stable and is likely to be
replaced by a different mechanism in a forthcoming version of this package.
}

\examples{
\dontrun{.add.extension('foo', foo.reader)}

}
\seealso{
\link{preinstalled.readers}
}
