\name{cache.project}
\alias{cache.project}
\title{
Cache all data sets that were loaded from the data directory.
}
\description{
Cache all data sets that were loaded from the data directory.
}
\usage{
cache.project()
}
\value{
	No value is returned; this function is called for its side effects.
}
\examples{
\dontrun{
library('ProjectTemplate')
load.project()

cache.project()}}
