% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.config.R
\name{create.config}
\alias{create.config}
\title{Add project specific config to the global config}
\usage{
create.config(...)
}
\arguments{
\item{...}{A series of key-value pairs containing the configuration.  The key is the
name that gets added to the config object.}
}
\description{
Enables project specific configuation to be added to the global config object
}
\examples{
library('ProjectTemplate')
\dontrun{
create.config(
    keep_bigdata=TRUE,     # Whether to keep the big data file in memory
    parse=7                # number of fields to parse 
)
}
}

