% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{data_generation}
\alias{data_generation}
\title{Data Generation Function}
\usage{
data_generation(iter = 10000, n = 1000, mu_x = 23, sd_x = 10,
  mu_y = 9, sd_y = 15,
  cores = as.numeric(Sys.getenv("LSB_DJOB_NUMPROC")))
}
\arguments{
\item{iter}{number of iterations to run}

\item{n}{number of observations in a single iteration to generate for x and y}

\item{mu_x}{mean for x variable passed to rnorm}

\item{sd_x}{standard deviation for x variable passed to rnorm}

\item{mu_y}{mean for y variable passed to rnorm}

\item{sd_y}{standard deviation for y variable passed to rnorm}

\item{cores}{number of cores to run it on}
}
\description{
Data Generation Function
}
\examples{
\dontrun{
library(broom)
library(parallel)
library(stringr)
library(tibble)
Run the function
data_generation(iter = 10000,
                n = 1000,
                mu_x = 23,
                sd_x = 10,
                mu_y = 9,
                sd_y = 15,
                cores = 1)
}
}
