% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preinstalled.readers.R
\docType{data}
\name{data_reader}
\alias{data_reader}
\title{Class to store reader in the dispatch table}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
data_reader
}
\description{
Normally you shouldn't need to construct an object of this class. Should you
need to do so, call \code{data_reader$new(name, reader)}.
}
\section{Fields}{

\describe{
\item{\code{name}}{Display name for the reader}

\item{\code{reader}}{Closure to import data}
}}

\keyword{internal}
