% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{default.config}
\alias{default.config}
\title{Default configuration}
\format{\preformatted{List of 9
 $ data_loading     : chr "on"
 $ cache_loading    : chr "on"
 $ recursive_loading: chr "off"
 $ munging          : chr "on"
 $ logging          : chr "off"
 $ load_libraries   : chr "off"
 $ libraries        : chr "reshape, plyr, ggplot2, stringr, lubridate"
 $ as_factors       : chr "on"
 $ data_tables      : chr "off"
}}
\usage{
default.config
}
\description{
This list stores the configuration used for new projects or as default
for missing items in the configuration of the current project.
}
\keyword{datasets}

