% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get.project}
\alias{get.project}
\title{Show information about the current project.}
\usage{
get.project(envir = .GlobalEnv)
}
\arguments{
  \item{envir}{The environment, defaults to the global
  environment.  In most use cases this parameter can be
  omitted.}
}
\value{
A named list.
}
\description{
This function will return all of the information that
ProjectTemplate has about the current project. This information is
gathered when \code{\link{load.project}} is called. At present,
ProjectTemplate keeps a record of the project's configuration settings,
all packages that were loaded automatically and all of the data sets that
were loaded automatically. The information about autoloaded data sets
is used by the \code{\link{cache.project}} function.
}
\examples{
library('ProjectTemplate')

\dontrun{load.project()

get.project()}
}
\seealso{
\code{\link{create.project}}, \code{\link{load.project}},
  \code{\link{cache.project}}, \code{\link{show.project}}
}

