% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.package.R
\name{.attach.or.add.namespace}
\alias{.attach.or.add.namespace}
\title{Attach a package or add a namespace}
\usage{
.attach.or.add.namespace(package.name, attach)
}
\arguments{
\item{package.name}{name of the package to load, as a character vector}

\item{attach}{boolean indicating whether to attach the package in the global namespace}
}
\value{
Boolean indicating whether the package was successfully loaded
}
\description{
Internal method to attach a package or only add the namespace.
}
\keyword{internal}
