% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cache.file.ext}
\alias{.cache.file.ext}
\title{Get file extensions of the cache files}
\usage{
.cache.file.ext(regex = FALSE)
}
\arguments{
\item{regex}{When \code{TRUE}, the file extensions are returned as regex
patterns.}
}
\value{
A named character vector containing the extensions of the data and
hash file.
}
\description{
Get file extensions of the cache files
}
\keyword{internal}
