% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cache.filename}
\alias{.cache.filename}
\title{Construct the file names for the cache and hash}
\usage{
.cache.filename(variable)
}
\arguments{
\item{variable}{Variable name for which to construct file names}
}
\value{
A list with file names
}
\description{
Construct the file names for the cache and hash
}
\details{
The returned object is a list with two fields:
\itemize{
  \item \code{obj}: The path to the file in which the
    variable contents will be saved;
  \item \code{hash}: The path to the file in which the cache
    metadata will be stored.
}
}
\keyword{internal}
