% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cache.format}
\alias{.cache.format}
\title{Get file extensions of the cache files}
\usage{
.cache.format()
}
\value{
A named character vector containing the extensions of the data and
hash file.
}
\description{
Get file extensions of the cache files
}
\keyword{internal}
