% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cache.hash}
\alias{.cache.hash}
\title{Calculate the hash of the data stored in a variable}
\usage{
.cache.hash(variables, env = .TargetEnv)
}
\arguments{
\item{variables}{character vector of variable names}

\item{env}{environment from which to load the variable}
}
\value{
data.frame with the variable names and the corresponding hashes
}
\description{
Calculate the hash of the data stored in a variable
}
\details{
The hashes are calculated using the \code{digest::digest} function.
}
\keyword{internal}
