% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cache.status}
\alias{.cache.status}
\title{Print the current cache status}
\usage{
.cache.status()
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
Print the current cache status
}
\keyword{internal}
