% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.project.R
\name{.check.version}
\alias{.check.version}
\title{Compare the project version with the current ProjectTemplate version}
\usage{
.check.version(config, warn.migrate = TRUE)
}
\arguments{
\item{config}{Project configuration}

\item{warn.migrate}{Logical indicating whether a warning should be raised if
the project version is older than the installed version of ProjectTemplate.}
}
\value{
\code{0} if the numbers are equal, \code{-1} if \code{b} is later
  and \code{1} if \code{a} is later (analogous to the C function
  \code{strcmp}).

  Gives an \R error on malformed inputs.
}
\description{
Compare the project version with the current ProjectTemplate version
}
\keyword{internal}
