% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.create.cache.hash}
\alias{.create.cache.hash}
\title{Create a data.frame with the cache metadata}
\usage{
.create.cache.hash(variable, depends, CODE)
}
\arguments{
\item{variable}{Name of the variable to be cached}

\item{depends}{Vector of variable names of dependencies for the variable to be cached, optional.}

\item{CODE}{Code block to generate \code{variable}, registered as a dependency, optional.}
}
\value{
\code{data.frame} containing the variable name and its dependencies, with the
  corresponding hashes appended.
}
\description{
Create a data.frame with the cache metadata
}
\details{
The hashes for the various objects are calculated using the \code{\link{.cache.hash}}
  function.
}
\seealso{
\code{\link{.cache.hash}}
}
\keyword{internal}
