% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.project.R
\name{.dir.empty}
\alias{.dir.empty}
\title{Check if a directory is empty}
\usage{
.dir.empty(path)
}
\arguments{
\item{path}{Character vector containing the path to the directory to check.}
}
\value{
Logical indicating whether the passed directory was empty.
}
\description{
Checks if the directory listing by \code{\link{.list.files.and.dirs}} is empty.
}
\keyword{file}
\keyword{internal}
