% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.evaluate.code}
\alias{.evaluate.code}
\title{Run code and assign the results to variable}
\usage{
.evaluate.code(variable, CODE)
}
\arguments{
\item{variable}{variable name in which to store the result of \code{CODE}}

\item{CODE}{code block that returns a result which can be stored in a variable}
}
\description{
Run code and assign the results to variable
}
\details{
No error handling is done on the executed code, nor is the
}
\keyword{internal}
