% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.template.R
\name{.get.template}
\alias{.get.template}
\title{Get the location of a template from its name}
\usage{
.get.template(template)
}
\arguments{
\item{template}{Character vector containing the name of the template}
}
\value{
Character vector containing the location of the template. If no
  template was found by the given name an error is raised.
}
\description{
Checks the configured \code{option('ProjectTemplate.templatedir')} for the
template. If no matching template was found the system templates are checked,
and finally the current directory is checked. If no template was found with
the given name an error is raised.
}
\keyword{internal}
