% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.project.R
\name{.has.project}
\alias{.has.project}
\title{Check if the project was loaded}
\usage{
.has.project()
}
\value{
Logical indicating whether the project was loaded.
}
\description{
Currently does a very basic check to see if the variable \code{project.info}
exists in the \code{.TargetEnv}. No check is performed on the contents of the
variable.
}
\keyword{internal}
