% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.project.R
\name{.init.logger}
\alias{.init.logger}
\title{Initialize the logger for the project}
\usage{
.init.logger(config, my.project.info)
}
\arguments{
\item{config}{Named list containing the project configuration}

\item{my.project.info}{Named list containing the project information}
}
\value{
Returns \code{my.project.info} amended with the new information.
}
\description{
Creates a \code{log4r::logger} and provides a default log file
\code{log/project.log}.
}
\keyword{internal}
