% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.is.cache.empty}
\alias{.is.cache.empty}
\title{Check whether the cache is empty}
\usage{
.is.cache.empty()
}
\value{
Logical indicating whether the cache is empty
}
\description{
Check whether the cache is empty
}
\keyword{internal}
