% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.is.cached}
\alias{.is.cached}
\title{Check whether variables are cached}
\usage{
.is.cached(varnames)
}
\arguments{
\item{varnames}{Character vector of variable names}
}
\value{
Logical vector indicating whether the variable is in the cache.
}
\description{
Check whether variables are cached
}
\keyword{internal}
