% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.project.R
\name{.is.dir}
\alias{.is.dir}
\title{Check if path is an existing directory}
\usage{
.is.dir(path)
}
\arguments{
\item{path}{Character vector containing the path to the directory to check.}
}
\value{
Logical indicating a valid directory was passed.
}
\description{
Checks if a given path exists, and if so if it is a directory.
}
\keyword{file}
\keyword{internal}
