% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopifnotproject.R
\name{.is.ProjectTemplate}
\alias{.is.ProjectTemplate}
\title{Test whether a given path is a ProjectTemplate project}
\usage{
.is.ProjectTemplate(path = getwd())
}
\arguments{
\item{path}{Directory to check, defaults to the current working directory.}
}
\value{
Logical indicating whether the given path is a valid project.
}
\description{
Test whether a given path is a ProjectTemplate project
}
\keyword{internal}
