% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.project.R
\name{.list.files.and.dirs}
\alias{.list.files.and.dirs}
\title{List all files and directories, excluding .. and .}
\usage{
.list.files.and.dirs(path)
}
\arguments{
\item{path}{Character vector indicating the path to the parent folder of
which the contents should be listed.}
}
\value{
Directory listing of \code{path}
}
\description{
Creates a directory listing of a given path, including hidden files and
subdirectories, but excluding the .. and . aliases.
}
\keyword{deprecate}
\keyword{file}
\keyword{internal}
