% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.project.R
\name{.load.data}
\alias{.load.data}
\title{Load the data from the cache and data directories}
\usage{
.load.data(config, my.project.info)
}
\arguments{
\item{config}{Named list containing the project configuration}

\item{my.project.info}{Named list containing the project information}
}
\value{
Returns \code{my.project.info} amended with the new information.
}
\description{
Gets the list of available variables in \code{cache/} and \code{data/} and
loads the data in memory. Data from the cache is loaded first, then in
alphabetical order.
}
\keyword{internal}
