% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.project.R
\name{.load.helpers}
\alias{.load.helpers}
\title{Load the helper functions}
\usage{
.load.helpers(config, my.project.info)
}
\arguments{
\item{config}{Named list containing the project configuration}

\item{my.project.info}{Named list containing the project information}
}
\value{
Returns \code{my.project.info} amended with the new information.
}
\description{
Sources all helper scripts in \code{lib}. If \code{lib/globals.R} exists this
is loaded first, all other scripts are sourced in alphabetical order.
}
\keyword{internal}
