% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.project.R
\name{.load.libraries}
\alias{.load.libraries}
\title{Load the libraries listed in the configuration into memory}
\usage{
.load.libraries(config, my.project.info)
}
\arguments{
\item{config}{Named list containing the project configuration}

\item{my.project.info}{Named list containing the project information}
}
\value{
Returns \code{my.project.info} amended with the new information.
}
\description{
Load the libraries listed in the libraries entry in global.dcf and add the
library names to the \code{project.info}.
}
\keyword{internal}
