% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.project.R
\name{.munge.data}
\alias{.munge.data}
\title{Source all munge scripts}
\usage{
.munge.data(config, my.project.info)
}
\arguments{
\item{config}{Named list containing the project configuration}

\item{my.project.info}{Named list containing the project information}
}
\value{
Returns \code{my.project.info} amended with the new information.
}
\description{
Sources all munge scripts in the \code{munge} directory in alphabetical
order.
}
\keyword{internal}
