% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.project.R
\name{.package.version}
\alias{.package.version}
\title{Get the current ProjectTemplate version}
\usage{
.package.version()
}
\value{
Version as a character vector.
}
\description{
Reads the installed version of ProjectTemplate from the \code{DESCRIPTION}
file.
}
\keyword{internal}
