% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.data.R
\name{.parse.extensions}
\alias{.parse.extensions}
\title{Match readers to the extensions of the data files}
\usage{
.parse.extensions(data.files)
}
\arguments{
\item{data.files}{a vector of paths to data files}
}
\value{
A list of \code{readers} and \code{varnames}
}
\description{
Match readers to the extensions of the data files
}
\keyword{internal}
