% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.data.R
\name{.prepare.data.ignore.regex}
\alias{.prepare.data.ignore.regex}
\title{Prepare a regular expression for matching files to be ignored}
\usage{
.prepare.data.ignore.regex(ignore_files)
}
\arguments{
\item{ignore_files}{A comma separated character vector that lists all
patterns to be matched for ignoring}
}
\value{
A chained regular expression that matches all patterns in the
  \code{ignore_files} variable.
}
\description{
Constructs a single regular expression for matching file names in data that
should not be imported. It can detect literal names, globs with wildcards and
regular expressions.
}
\keyword{internal}
