% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.project.R
\name{.provide.directory}
\alias{.provide.directory}
\title{Make sure a required directory exists before usage}
\usage{
.provide.directory(name)
}
\arguments{
\item{name}{Character vector containing the name of the required directory.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
Checks if the requested directory exists, and if not creates the directory.
In the latter case a warning is raised.
}
\keyword{internal}
