% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopifnotproject.R
\name{.quietstop}
\alias{.quietstop}
\title{Stop silently}
\usage{
.quietstop()
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
Temporarily disable \code{option(show.error.messages)} and stop execution.
}
\keyword{deprecate}
\keyword{internal}
