% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.read.cache.info}
\alias{.read.cache.info}
\title{Read metadata for a variable in the cache}
\usage{
.read.cache.info(variable)
}
\arguments{
\item{variable}{Variable name for which to look up the metadata}
}
\value{
\code{list} with metadata, see Details for more info.
}
\description{
Read metadata for a variable in the cache
}
\details{
The returned object is a list with two fields:

\itemize{
  \item \code{in.cache}: Logical indicating whether the requested
    variable was found in the cache
  \item \code{hash}: A data.frame as was created by \code{\link{.create.cache.hash}}
}
}
\keyword{internal}
