% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{.remove.sticky.vars}
\alias{.remove.sticky.vars}
\title{Remove variables to keep from a list of candidates for removal}
\usage{
.remove.sticky.vars(names, keep)
}
\arguments{
\item{names}{character vector of variable names that are candidate for removal}

\item{keep}{character vector of variable names that should not be removed}
}
\value{
A character vector containing the variables to remove.
}
\description{
Remove variables to keep from a list of candidates for removal
}
\details{
If the \code{sticky_variables} option is part of the \code{config}
  variable the \code{config} variable itself is added to the list of variables
  to keep. Also all variables listed in \code{config$sticky_variables} in a
  comma separated list are added to keep.
}
\keyword{internal}
