% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.package.R
\name{.require.package}
\alias{.require.package}
\title{Require internal package}
\usage{
.require.package(package.name)
}
\arguments{
\item{package.name}{name of the package to load, as a character vector}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
Internal method to require a package that is necessary for the internal
functioning of ProjectTemplate. Never attaches the package unless
configured to do so in global.dcf (which throws a warning).
}
\keyword{internal}
