% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.project.R
\name{.rstudioprojectfile}
\alias{.rstudioprojectfile}
\title{Return an RStudio project file as character vector}
\usage{
.rstudioprojectfile()
}
\value{
Character vector with the contents of an empty RStudio project file
}
\description{
Return an RStudio project file as character vector
}
\keyword{internal}
