% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopifnotproject.R
\name{.stopifnotproject}
\alias{.stopifnotproject}
\title{Raise an error if given path is not a valid project}
\usage{
.stopifnotproject(additional_message = "", path = getwd())
}
\arguments{
\item{additional_message}{Optional message to show if the given path is not a
valid project}

\item{path}{Path to check if it is a valid project}
}
\value{
Project name if it is a valid Project.
}
\description{
Function to stop processing if the path is not a Project Template return the
project name if it is a Project Template directory.
}
\keyword{internal}
