% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopifnotproject.R
\name{.stopifproject}
\alias{.stopifproject}
\title{Raise an error if given path is a valid project}
\usage{
.stopifproject(additional_message = "", path = getwd())
}
\arguments{
\item{additional_message}{Optional message to show if the given path is not a
valid project}

\item{path}{Path to check if it is a valid project}
}
\value{
No value is returned; this function is called for its side effects
}
\description{
Function to stop processing if the path is a Project Template.
}
\keyword{internal}
