% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.project.R
\name{.unload.project}
\alias{.unload.project}
\title{Unload the project variables keeping the data}
\usage{
.unload.project()
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
Removes the \code{config}, \code{logger} and \code{project.info} variables
from memory, leaving all data variables in place.
}
\keyword{deprecate}
\keyword{internal}
