% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.project.R
\name{.var.diff.from}
\alias{.var.diff.from}
\title{Compare sets of variable names}
\usage{
.var.diff.from(given.var.list = "", env = .TargetEnv)
}
\arguments{
\item{given.var.list}{Character vector of variable names}

\item{env}{Environment in which to compare the sets of variables}
}
\description{
Compare the variables (excluding functions) in the global env with a passed
in string of names and return the set difference.
}
\keyword{internal}
