% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.write.cache}
\alias{.write.cache}
\title{Write a variable and its metadata to cache}
\usage{
.write.cache(cache.hash, ...)
}
\arguments{
\item{cache.hash}{a \code{data.frame} with metadata about the variable, see details for more information.}

\item{...}{extra parameters passed to \code{\link{save}}.}
}
\value{
No value is returned, this function is called for its side effects.
}
\description{
Write a variable and its metadata to cache
}
\details{
cache.hash is a data frame with two columns: \code{variable} and \code{hash}.\cr
  Row name \code{VAR} is the name of the variable to save.\cr
  Row name \code{CODE} is the hash value of the code to compute variable.\cr
  Row name \code{DEPENDS.*} are the dependent variables that \code{CODE} depends on.c\cr
  The helper function \code{\link{.create.cache.hash}} creates a suitable dataframe
}
\keyword{internal}
