% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{is.cached}
\alias{is.cached}
\title{Check whether a variable is in the cache}
\usage{
is.cached(variable)
}
\arguments{
\item{variable}{A character string containing the name of the variable to
be checked.}
}
\value{
\code{TRUE} if the variable exists in the cache, \code{FALSE} otherwise
}
\description{
This function will determine if a variable is stored in the \code{cache}
directory.
}
\examples{
library('ProjectTemplate')
\dontrun{is.cached('mapdata')
}
}

