% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.data.R
\name{list.data}
\alias{list.data}
\title{Listing the data for the current project}
\usage{
list.data(override.config = NULL)
}
\arguments{
\item{override.config}{Named list, allows overriding individual configuration
items.}
}
\value{
df data.frame containing available data with relevant meta data
}
\description{
This function produces a data.frame of all data files in the project, with
  the following meta data:
}
\examples{
library('ProjectTemplate')

\dontrun{list.data()}
}
\seealso{
\code{\link{load.project}}, \code{\link{show.project}},
   \code{\link{project.config}}
}

