% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{loadproject_addin}
\alias{loadproject_addin}
\title{Load Project}
\usage{
loadproject_addin()
}
\description{
Call this function as an addin to load the library and run `load.project()`
}
\keyword{internal}
