% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{migrate.project}
\alias{migrate.project}
\title{Migrates a project from a previous version of ProjectTemplate}
\usage{
migrate.project()
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function automatically performs all necessary steps to migrate an existing project
so that it is compatible with this version of ProjectTemplate
}
\examples{
library('ProjectTemplate')

\dontrun{migrate.project()}
}
\seealso{
\code{\link{create.project}}
}

