% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.template.R
\name{migrate.template}
\alias{migrate.template}
\title{Migrate a template to a new version of ProjectTemplate}
\usage{
migrate.template(template)
}
\arguments{
\item{template}{Name of the template to upgrade.}
}
\description{
This function updates a skeleton project to the current version of
  ProjectTemplate.
}
