% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{data}
\name{new.config}
\alias{new.config}
\title{Configuration for new projects}
\format{\preformatted{List of 11
 $ version                  : chr "0.6"
 $ data_loading             : chr "TRUE"
 $ cache_loading            : chr "TRUE"
 $ recursive_loading        : chr "FALSE"
 $ munging                  : chr "TRUE"
 $ logging                  : chr "FALSE"
 $ load_libraries           : chr "FALSE"
 $ libraries                : chr "reshape, plyr, ggplot2, stringr, lubridate"
 $ as_factors               : chr "TRUE"
 $ data_tables              : chr "FALSE"
 $ attach_internal_libraries: chr "FALSE"
}}
\usage{
new.config
}
\description{
This list stores the configuration used for new projects.
}
\keyword{datasets}

