% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{data}
\name{preinstalled.readers}
\alias{preinstalled.readers}
\title{Maps file types to the reader functions used to autoload them.}
\format{\preformatted{List of 52
 $ *.csv     : chr "csv.reader"
 $ *.csv.bz2 : chr "csv.reader"
 $ *.csv.zip : chr "csv.reader"
 $ *.csv.gz  : chr "csv.reader"
 $ *.csv2    : chr "csv2.reader"
 $ *.csv2.bz2: chr "csv2.reader"
 $ *.csv2.zip: chr "csv2.reader"
 $ *.csv2.gz : chr "csv2.reader"
 $ *.tsv     : chr "tsv.reader"
 $ *.tsv.bz2 : chr "tsv.reader"
 $ *.tsv.zip : chr "tsv.reader"
 $ *.tsv.gz  : chr "tsv.reader"
 $ *.tab     : chr "tsv.reader"
 $ *.tab.bz2 : chr "tsv.reader"
 $ *.tab.zip : chr "tsv.reader"
 $ *.tab.gz  : chr "tsv.reader"
 $ *.wsv     : chr "wsv.reader"
 $ *.wsv.bz2 : chr "wsv.reader"
 $ *.wsv.zip : chr "wsv.reader"
 $ *.wsv.gz  : chr "wsv.reader"
 $ *.txt     : chr "wsv.reader"
 $ *.txt.bz2 : chr "wsv.reader"
 $ *.txt.zip : chr "wsv.reader"
 $ *.txt.gz  : chr "wsv.reader"
 $ *.Rdata   : chr "rdata.reader"
 $ *.rda     : chr "rdata.reader"
 $ *.R       : chr "r.reader"
 $ *.r       : chr "r.reader"
 $ *.url     : chr "url.reader"
 $ *.sql     : chr "sql.reader"
 $ *.xls     : chr "xls.reader"
 $ *.xlsx    : chr "xlsx.reader"
 $ *.sav     : chr "spss.reader"
 $ *.dta     : chr "stata.reader"
 $ *.arff    : chr "arff.reader"
 $ *.dbf     : chr "dbf.reader"
 $ *.rec     : chr "epiinfo.reader"
 $ *.mtp     : chr "mtp.reader"
 $ *.m       : chr "octave.reader"
 $ *.sys     : chr "systat.reader"
 $ *.syd     : chr "systat.reader"
 $ *.sas     : chr "xport.reader"
 $ *.xport   : chr "xport.reader"
 $ *.xpt     : chr "xport.reader"
 $ *.db      : chr "db.reader"
 $ *.file    : chr "file.reader"
 $ *.mp3     : chr "mp3.reader"
 $ *.ppm     : chr "ppm.reader"
 $ *.dat     : chr "wsv.reader"
 $ *.dat.bz2 : chr "wsv.reader"
 $ *.dat.zip : chr "wsv.reader"
 $ *.dat.gz  : chr "wsv.reader"
}}
\usage{
preinstalled.readers
}
\description{
This list stores a mapping from regular expressions that match file
extensions for the file types supported by ProjectTemplate to the
reader functions that implement autoloading for those formats. Any
new file type must be appended to this dispatch table.
}
\keyword{datasets}

