% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.config.R
\name{project.config}
\alias{project.config}
\title{ProjectTemplate Configuration file}
\usage{
project.config()
}
\arguments{
\item{data_loading}{This can be set to 'on' or 'off'. If data_loading is on,
the system will load data from both the cache and data directories with 
cache taking precedence in the case of name conflict.}

\item{cache_loading}{This can be set to 'on' or 'off'. If cache_loading is on,
the system will load data from the cache directory before any attempt to load 
from the data directory.}

\item{recursive_loading}{This can be set to 'on' or 'off'. If recursive_loading
is on, the system will load data from the data directory and all its sub 
directories recursively.}

\item{munging}{This can be set to 'on' or 'off'. If munging is on, the system
will execute the files in the munge directory sequentially using the order 
implied by the sort() function. If munging is off, none of the files in the
 munge directory will be executed.}

\item{logging}{This can be set to 'on' or 'off'. If logging is on, a logger 
object using the log4r package is automatically created when you run 
load.project(). This logger will write to the logs directory.}

\item{logging_level}{The value of logging_level is passed to  a logger object 
using the log4r package during logging when when you run load.project().}

\item{load_libraries}{This can be set to 'on' or 'off'. If load_libraries is on,
the system will load all of the R packages listed in the libraries field 
described below.}

\item{libraries}{This is a comma separated list of all the R packages that the user
wants to automatically load when load.project() is called. These packages must 
already be installed before calling load.project().}

\item{as_factors}{This can be set to 'on' or 'off'. If as_factors is on, the system
will convert every character vector into a factor when creating data frames; most 
importantly, this automatic conversion occurs when reading in data automatically. 
If 'off', character vectors will remain character vectors.}

\item{data_tables}{This can be set to 'on' or 'off'. If data_tables is on, the 
system will convert every data set loaded from the data directory into a data.table.}

\item{attach_internal_libraries}{This can be set to 'on' or 'off'. If 
attach_internal_libraries is on, then every time a new package is loaded into memory 
during load.project() a warning will be displayed informing that has happened.}

\item{cache_loaded_data}{This can be set to 'on' or 'off'. If cache_loaded_data is
on, then data loaded from the data directory during load.project() will be 
automatically cached (so it won't need to be reloaded next time load.project() 
is called).}
}
\value{
The current project configuration is displayed.
}
\description{
Every \code{ProjectTemplate} project has a configuration file found at
\code{config/global.dcf} that contains various options that can be tweaked
to control runtime behaviour.  The valid options are shown below, and must
be encoded using the \code{DCF} format.
}
\details{
Calling the \code{project.config()} function will display the current project
configuration.

 If the \code{config/globals.dcf} is missing some items (for example because it was created under an
old version of \code{ProjectTemplate}, then the following configuration is used for any missing items
during \code{load.project()}:
  \Sexpr[results=rd, stage=build]{ProjectTemplate:::.format.config(ProjectTemplate:::.default.config[-1], format="Rd")}

 When a new project is created using \code{create.project()}, the following values are pre-populated:
  \Sexpr[results=rd, stage=build]{ProjectTemplate:::.format.config(ProjectTemplate:::.new.config, format="Rd")}
}
\seealso{
\code{\link{load.project}}
}

