\name{ProjectTemplate}
\title{Automates the creation of new statistical analysis projects.}

\description{
  Automates the creation of new statistical analysis
  projects.
}

\details{
  ProjectTemplate provides functions to automatically build
  a directory structure for a new R project. Using this
  structure, ProjectTemplate automates data loading,
  preprocessing, library importing and unit testing.
}
\references{This code is inspired by the skeleton structure used by
Ruby on Rails.}
\docType{package}
\alias{ProjectTemplate}
\alias{package-ProjectTemplate}
\examples{library('ProjectTemplate')

#create.project('project_name')

#setwd('project_name')
#load.project()}
