% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.reader.R
\name{rds.reader}
\alias{rds.reader}
\title{Read the RDS file format.}
\usage{
rds.reader(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load a data set stored in the RDS file format into
the specified global variable binding.
}
\examples{
\dontrun{rds.reader('example.rds', 'data/example.rds', 'example')}
}
