% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{reloadproject_addin}
\alias{reloadproject_addin}
\title{Reload Project}
\usage{
reloadproject_addin()
}
\description{
Call this function as an addin to load the library and run `reload.project()`
}
\keyword{internal}
