\name{run.tests}
\alias{run.tests}
\title{
Automatically run all of the tests cases in the tests directory.
}
\description{
Automatically run all of the tests cases in the tests directory.
}
\usage{
run.tests()
}
\arguments{
}
}
\value{
	No value is returned; this function is called for its side effects.
}
\examples{
library('ProjectTemplate')

create.project('project_name')

setwd('project_name')

run.test()}
