\name{show.project}
\alias{show.project}
\title{
Show the current project's internal state.
}
\description{
Show the current project's internal state.
}
\usage{
show.project()
}
\value{
	No value is returned; this function is called for its side effects.
}
\examples{
\dontrun{
library('ProjectTemplate')
load.project()

show.project()}}
