\name{show.updates}
\alias{show.updates}
\title{
Show the user how their directory structure would be different under the most recent version of ProjectTemplate.
}
\description{
Show the user how their directory structure would be different under the most recent version of ProjectTemplate. Currently uses diff and will therefore not run on Windows systems.
}
\usage{
show.updates()
}
\value{
	No value is returned; this function is called for its side effects.
}
\examples{
\dontrun{
library('ProjectTemplate')

show.updates()}}

