\name{stub.tests}
\alias{stub.tests}
\title{
Autogenerate tests for helper function.
}
\description{
Automatically generates one trivial test for every helper function found in the lib directory. The results are stored in tests/autogenerated.R.
}
\usage{
stub.tests()
}
\value{
	No value is returned; this function is called for its side effects.
}
\examples{
\dontrun{
library('ProjectTemplate')

create.project('project_name')

setwd('project_name')

stub.tests()}}
