% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preinstalled.readers.R
\name{type_sum.data_reader}
\alias{type_sum.data_reader}
\title{Display the reader name in tibbles}
\usage{
type_sum.data_reader(x)
}
\arguments{
\item{x}{reader for which to show the type summary}
}
\description{
Display the reader name in tibbles
}
\keyword{internal}
