% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.extension.R
\name{.add.extension}
\alias{.add.extension}
\title{Associate a reader function with an extension.}
\usage{
.add.extension(extension, reader)
}
\arguments{
\item{extension}{The extension of the new data file.}

\item{reader}{The function to use when reading the data file.  It should
accept three arguments: \code{data.file}, \code{filename} and
\code{variable.name} (in that order).  The function should read the
contents of the file \code{filename}, and save it into the workspace
under the name \code{variable.name}.  The \code{data.file} argument
is just a relative file name and can be ignored.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will associate an extension with a custom reader function.
}
\section{Warning}{

This interface should not be considered as stable and is likely to be
replaced by a different mechanism in a forthcoming version of this package.
}
\examples{
\dontrun{.add.extension('foo', foo.reader)}
}

