% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.project.R
\name{cache.project}
\alias{cache.project}
\title{Cache a project's data sets in binary format.}
\usage{
cache.project()
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will cache all of the data sets that were loaded by
the \code{\link{load.project}} function in a binary format that is
easier to load quickly. This is particularly useful for data sets
that you've modified during a slow munging process that does not
need to be repeated.
}
\examples{
library('ProjectTemplate')
\dontrun{load.project()

cache.project()}
}
\seealso{
\code{\link{create.project}}, \code{\link{load.project}},
  \code{\link{get.project}}, \code{\link{show.project}}
}

