% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.variable.name.R
\name{clean.variable.name}
\alias{clean.variable.name}
\title{Translate a file name into a valid R variable name.}
\usage{
clean.variable.name(variable.name)
}
\arguments{
\item{variable.name}{A character vector containing a variable's proposed
name that should be standardized.}
}
\value{
A translated variable name.
}
\description{
This function will translate a file name into a name that is a valid
variable name in R. Non-alphabetic characters on the boundaries of the
file name will be stripped; non-alphabetic characters inside of the file
name will be replaced with dots.
}
\examples{
library('ProjectTemplate')

\dontrun{clean.variable.name('example_1')}
}

