% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.cache.R
\name{clear.cache}
\alias{clear.cache}
\title{Clear data sets from the cache.}
\usage{
clear.cache(variables = NULL)
}
\arguments{
\item{variables}{A character vector containing the name of the variable to
be removed from the cache.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function remove specific (or all by default) named data sets from the \code{cache}
directory. This will force that data to be read in from the \code{data} directory
next time \code{\link{load.project}} is called.
}
\examples{
library('ProjectTemplate')
\dontrun{clear.cache()}


}

