% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2.reader.R
\name{csv2.reader}
\alias{csv2.reader}
\title{Read a semicolon separated values (.csv2) file.}
\usage{
csv2.reader(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load a data set stored in the CSV2 file format into
the specified global variable binding.
}
\examples{
library('ProjectTemplate')

\dontrun{csv2.reader('example.csv2', 'data/example.csv2', 'example')}
}

