% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.config.R
\docType{data}
\name{default.config}
\alias{default.config}
\title{Default configuration}
\format{An object of class \code{list} of length 11.}
\usage{
default.config
}
\description{
This list stores the configuration used for missing items
in the configuration of the current project.
}
\keyword{datasets}

