\name{extensions.dispatch.table}
\alias{extensions.dispatch.table}
\title{Maps file types to the reader functions used to autoload them.}
\description{
  This list stores a mapping from regular expressions that
  match file extensions for the file types supported by
  ProjectTemplate to the reader functions that implement
  autoloading for those formats. Any new file type must be
  appended to this dispatch table.
}

