% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.cache.format}
\alias{.cache.format}
\title{Get configured cache file format strategy}
\usage{
.cache.format()
}
\value{
A named object of mode \code{\link{expression}}.
}
\description{
Get configured cache file format strategy
}
\keyword{internal}
