\name{load.project}
\alias{load.project}
\title{Automatically load data and packages for a project.}

\description{
  Automatically load data and packages for a project.
}

\details{
  This function automatically load all of the data and
  packages used by the project from which it is called.
}
\value{No value is returned; this function is called for its side effects.}
\examples{library('ProjectTemplate')

#load.project()}
