% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.project.R
\name{migrate.project}
\alias{migrate.project}
\title{Migrates a project from a previous version of ProjectTemplate}
\usage{
migrate.project()
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function automatically performs all necessary steps to migrate an existing project
so that it is compatible with this version of ProjectTemplate
}
\examples{
library('ProjectTemplate')

\dontrun{migrate.project()}
}
\seealso{
\code{\link{create.project}}
}

