% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp.reader.R
\name{mtp.reader}
\alias{mtp.reader}
\title{Read a Minitab Portable Worksheet with an .mtp3 file extension.}
\usage{
mtp.reader(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified Minitab Portable Worksheet into
memory.
}
\examples{
library('ProjectTemplate')

\dontrun{mtp.reader('example.mtp', 'data/example.mtp', 'example')}
}

