% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.config.R
\docType{data}
\name{new.config}
\alias{new.config}
\title{Configuration for new projects}
\format{An object of class \code{list} of length 11.}
\usage{
new.config
}
\description{
This list stores the configuration used for new projects.
}
\keyword{datasets}

