% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octave.reader.R
\name{octave.reader}
\alias{octave.reader}
\title{Read an Octave file with a .m file extension.}
\usage{
octave.reader(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified Octave file into memory.
}
\examples{
library('ProjectTemplate')

\dontrun{octave.reader('example.m', 'data/example.m', 'example')}
}

