% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.reader.R
\name{ppm.reader}
\alias{ppm.reader}
\title{Read a PPM file with a .ppm file extension.}
\usage{
ppm.reader(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified PPM file into memory using the
pixamp package. This is useful for working with image files as a data
set.
}
\examples{
library('ProjectTemplate')

\dontrun{ppm.reader('example.ppm', 'data/example.ppm', 'example')}
}

