% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preinstalled.readers.R
\docType{data}
\name{preinstalled.readers}
\alias{preinstalled.readers}
\title{Maps file types to the reader functions used to autoload them.}
\format{An object of class \code{list} of length 52.}
\usage{
preinstalled.readers
}
\description{
This list stores a mapping from regular expressions that match file
extensions for the file types supported by ProjectTemplate to the
reader functions that implement autoloading for those formats. Any
new file type must be appended to this dispatch table.
}
\keyword{datasets}

