% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectTemplate-package.R
\docType{package}
\name{ProjectTemplate-package}
\alias{ProjectTemplate}
\alias{ProjectTemplate-package}
\title{ProjectTemplate: Automates the Creation of New Statistical Analysis Projects}
\description{
Provides functions to automatically build a directory structure for a new R project. Using this structure, 'ProjectTemplate' automates data loading, preprocessing, library importing and unit testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://projecttemplate.net}
  \item Report bugs at \url{https://github.com/KentonWhite/ProjectTemplate/issues}
}

}
\author{
\strong{Maintainer}: Kenton White \email{jkentonwhite@gmail.com} [contributor]

Authors:
\itemize{
  \item John Myles White [copyright holder]
}

Other contributors:
\itemize{
  \item Aleksandar Blagotic [contributor]
  \item Diego Valle-Jones [contributor]
  \item Jeffrey Breen [contributor]
  \item Joakim Lundborg [contributor]
  \item Josh Bode [contributor]
  \item Kirill Mueller [contributor]
  \item Matteo Redaelli [contributor]
  \item Noah Lorang [contributor]
  \item Patrick Schalk [contributor]
  \item Dominik Schneider [contributor]
  \item Gerold Hepp [contributor]
  \item Zunaira Jamile [contributor]
}

}
\keyword{internal}
