% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{ProjectTemplate}
\alias{ProjectTemplate}
\alias{ProjectTemplate-package}
\alias{package-ProjectTemplate}
\title{Automates the creation of new statistical analysis projects.}
\description{
ProjectTemplate provides functions to automatically build a directory
structure for a new R project. Using this structure, ProjectTemplate
automates data loading, preprocessing, library importing and unit
testing.
}
\examples{
library('ProjectTemplate')

\dontrun{create.project('project_name')

setwd('project_name')
load.project()}
}
\references{
This code is inspired by the skeleton structure used by
  Ruby on Rails.
}

