\name{reload.project}
\alias{reload.project}
\title{Reload a project from scratch.}

\description{
  Reload a project from scratch.
}

\details{
  This function will clear the global environment and
  reload a project from scratch. This is useful when you've
  updated your data sets or changed your preprocessing
  scripts.
}
\value{No value is returned; this function is called for its side effects.}
\examples{library('ProjectTemplate')

#load.project()

#reload.project()}
