% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separated.list.R
\name{separated.list}
\alias{separated.list}
\title{Create a concatenated string from an R list.}
\usage{
separated.list(sepchar = ";", target.list, ignore)
}
\arguments{
\item{sepchar}{this is the character with which you wish to separate list.}

\item{target.list}{this is a list to be concatenated. It is in the form
of the information translated from a .dcf file using
\code{\link{translate.dcf}}}

\item{ignore}{this is an optional character vector to ignore particular
elements in the list.}
}
\description{
This creates a list separated by the specified \code{sepchar}, which is
useful for specifying \code{\link{sql.reader}} odbc connections which
need a concatenated string separated by a semi-colon. It ignores
list elements that are blank.
It means that the \code{sql.reader} functions can be more flexible when
reading .sql dcf files, as the user can specify the inputs for
\code{\link[RODBC]{odbcDriverConnect}}.
}
\examples{
library('ProjectTemplate')

\dontrun{
 separated.list(sepchar = ";",
   target.list = database.info,
   ignore = c("ridiculousoption")
   )
   }

}
\seealso{
\code{\link[ProjectTemplate]{sql.reader}} for useage and
\code{\link[ProjectTemplate]{translate.dcf}} for \code{database.info} form.
}

