% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata.reader.R
\name{stata.reader}
\alias{stata.reader}
\title{Read a Stata file with a .stata file extension.}
\usage{
stata.reader(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified Stata file into memory.
}
\examples{
library('ProjectTemplate')

\dontrun{stata.reader('example.stata', 'data/example.stata', 'example')}
}

