\name{stub.tests}
\alias{stub.tests}
\title{Generate unit tests for your helper functions.}

\description{
  Generate unit tests for your helper functions.
}

\details{
  This function will parse all of the functions defined in
  files inside of the \code{lib} directory and will
  generate a trivial unit test for each function. The
  resulting tests are stored in the file
  \code{tests/autogenerated.R}. Every test is excepted to
  fail by default, so you should edit them before calling
  \code{\link{test.project}}.
}
\value{No value is returned; this function is called for its side effects.}
\examples{library('ProjectTemplate')

#stub.tests()}
