% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stub.tests.R
\name{stub.tests}
\alias{stub.tests}
\title{Generate unit tests for your helper functions.}
\usage{
stub.tests()
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will parse all of the functions defined in files inside
of the \code{lib} directory and will generate a trivial unit test for
each function. The resulting tests are stored in the file
\code{tests/autogenerated.R}. Every test is excepted to fail by default,
so you should edit them before calling \code{\link{test.project}}.
}
\examples{
library('ProjectTemplate')

\dontrun{stub.tests()}
}

