% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsv.reader.R
\name{tsv.reader}
\alias{tsv.reader}
\title{Read a tab separated values (.tsv or .tab) file.}
\usage{
tsv.reader(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load a data set stored in the TSV file format into
the specified global variable binding.
}
\examples{
library('ProjectTemplate')

\dontrun{tsv.reader('example.tsv', 'data/example.tsv', 'example')}
}

