% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xport.reader.R
\name{xport.reader}
\alias{xport.reader}
\title{Read an XPort file with a .xport file extension.}
\usage{
xport.reader(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified XPort file into memory.
}
\examples{
library('ProjectTemplate')

\dontrun{xport.reader('example.xport', 'data/example.xport', 'example')}
}

